nexus_etl/requirements.txt
pandas==1.3.1
pyxnat==1.4
PyYAML==6.0
rdflib==6.1.1
requests==2.24.0
SPARQLWrapper==2.0.0
SQLAlchemy==1.4.22
xmltodict==0.12.0
nexus-sdk @ git+https://github.com/BlueBrain/nexus-python-sdk
